Updates - Bug Fixes
Added render state checks to save up on render state calls when not needed in the DX9Backend implementation.
Removed D3D_DEBUG_INFO definition from the files and moved it as a global define in CMakeLists.
Changed binding order from pixel->vertex to vertex->pixel in DX9ShaderProgram.
Fixed an issue with matrix constant setter in DX9ShaderProgram.
Texture upload routine was modified to use ID3D9Device::CreateTexture instead of D3DXCreateTexture.
Added texture sampler state configuration in the DX9Texture implementation.
Fixed a logical error in DX9VertexBuffer that would continuously try to re-allocate the Vertex Buffer when the current buffer size is less than the new buffer size.
Added logic in DX9VertexBuffer to set the correct usage configuration for the buffer when the buffer is hinted to be streamed.